# frontend/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CRM System API",
    "description": "REST API for managing a foreign language school",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
...
    "/api/students/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "operationId": "GetStudentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Students"
        ],
        "operationId": "UpdateStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "operationId": "DeleteStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/students/{id}/preferences": {
      "put": {
        "tags": [
          "Students"
        ],
        "operationId": "UpdateStudentPreferences",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/teachers/me/students": {
      "get": {
        "tags": [
          "Teachers"
        ],
        "operationId": "GetMyStudents",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyStudentResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/teachers/{id}": {
      "get": {
        "tags": [
          "Teachers"
        ],
        "operationId": "GetTeacherById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Teachers"
        ],
        "operationId": "UpdateTeacher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Teachers"
        ],
        "operationId": "DeleteTeacher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Format": {
        "enum": [
          "Online",
          "Offline",
          null
        ]
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Language": {
        "enum": [
          "English",
          "French",
          "German",
          "Polish",
          "Spanish",
          "Italian",
          null
        ]
      },
      "LearningGoal": {
        "enum": [
          "Work",
          "Relocation",
          "Study",
          "Personal"
        ]
      },
...
      "LessonType": {
        "enum": [
          "Individual",
          "Group"
        ]
      },
      "Level": {
        "enum": [
          "A1",
          "A2",
          "B1",
          "B2",
          "C1",
          "C2",
          null
        ]
      },
...
      "MyStudentResponse": {
        "required": [
          "id",
          "fullName",
          "email",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
...
      "ParentInfoResponse": {
        "required": [
          "firstName",
          "lastName",
          "middleName",
          "phoneNumber",
          "email"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": [
              "null",
              "string"
            ]
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StudentDetailResponse": {
        "required": [
          "id",
          "firstName",
          "lastName",
          "middleName",
          "dateOfBirth",
          "phoneNumber",
          "email",
          "city",
          "country",
          "isChild",
          "comment",
          "status",
          "hasAccount",
          "preferences",
          "languages",
          "parentInfo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": [
              "null",
              "string"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "isChild": {
            "type": "boolean"
          },
          "comment": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/StudentStatus"
          },
          "hasAccount": {
            "type": "boolean"
          },
          "preferences": {
            "$ref": "#/components/schemas/StudentPreferencesResponse"
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            }
          },
          "parentInfo": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParentInfoResponse"
              }
            ]
          }
        }
      },
...
      "StudentPreferencesResponse": {
        "required": [
          "learningGoal",
          "format",
          "lessonType",
          "intensity",
          "currentLevel",
          "hadPreviousCourses"
        ],
        "type": "object",
        "properties": {
          "learningGoal": {
            "$ref": "#/components/schemas/LearningGoal"
          },
          "format": {
            "$ref": "#/components/schemas/Format"
          },
          "lessonType": {
            "$ref": "#/components/schemas/LessonType"
          },
          "intensity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currentLevel": {
            "$ref": "#/components/schemas/Level"
          },
          "hadPreviousCourses": {
            "type": "boolean"
          }
        }
      },
...
      "StudentStatus": {
        "enum": [
          "Active",
          "Suspended",
          "Graduated",
          "Withdrawn"
        ]
      },
...
      "TeacherDetailResponse": {
        "required": [
          "id",
          "userId",
          "firstName",
          "lastName",
          "middleName",
          "dateOfBirth",
          "phoneNumber",
          "email",
          "city",
          "country",
          "status",
          "comment",
          "hasAccount",
          "currentSalaryRate",
          "salaryRates"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": [
              "null",
              "string"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TeacherStatus"
          },
          "comment": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasAccount": {
            "type": "boolean"
          },
          "currentSalaryRate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TeacherSalaryRateResponse"
              }
            ]
          },
          "salaryRates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeacherSalaryRateResponse"
            }
          }
        }
      },
...
      "TeacherSalaryRateResponse": {
        "required": [
          "id",
          "baseSalary",
          "lessonsRate",
          "effectiveFrom"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "baseSalary": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "lessonsRate": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TeacherStatus": {
        "enum": [
          "Probation",
          "Employed",
          "OnLeave",
          "Resigned",
          "Dismissed"
        ]
      },
...
      "UpdatePreferencesRequest": {
        "required": [
          "learningGoal",
          "format",
          "lessonType",
          "intensity",
          "currentLevel",
          "hadPreviousCourses",
          "languages"
        ],
        "type": "object",
        "properties": {
          "learningGoal": {
            "$ref": "#/components/schemas/LearningGoal"
          },
          "format": {
            "$ref": "#/components/schemas/Format"
          },
          "lessonType": {
            "$ref": "#/components/schemas/LessonType"
          },
          "intensity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currentLevel": {
            "$ref": "#/components/schemas/Level"
          },
          "hadPreviousCourses": {
            "type": "boolean"
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            }
          }
        }
      },
...
      "UpdateRequest": {
        "required": [
          "firstName",
          "lastName",
          "middleName",
          "dateOfBirth",
          "phoneNumber",
          "city",
          "country",
          "isChild"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": [
              "null",
              "string"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "phoneNumber": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "isChild": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateTeacherRequest": {
        "required": [
          "firstName",
          "lastName",
          "middleName",
          "dateOfBirth",
          "phoneNumber",
          "city",
          "country"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": [
              "null",
              "string"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "phoneNumber": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}